fix: upload bug

+++ src/behaviors/UploadBehavior.php
@@ -34,17 +34,17 @@ class UploadBehavior extends \yii\base\Behavior
    {
 
       foreach ($this->owner->rules() as $rule) {
-         if($rule[1] === 'file' || $rule[1] === 'image') {
+         if ($rule[1] === 'file' || $rule[1] === 'image') {
             if (is_array($rule[0])) {
                foreach ($rule[0] as $attr) {
                   $thisAttr = $this->owner->getAttribute($attr);
                   $attrValues = [];
-                  if($thisAttr !== null && is_array($thisAttr)) {
+                  if ($thisAttr !== null && is_array($thisAttr)) {
                      foreach ($thisAttr as $thisFile) {
                         $newfile = tempnam(sys_get_temp_dir(), "_tmp");
                         $bytes = file_get_contents(str_replace(' ', '+', $thisFile['data']));
                         file_put_contents($newfile, $bytes);
-                        $attrValues[]= new UploadedFile(['name' => $thisFile['name'], 'tempName' => $newfile, 'type' => \yii\helpers\FileHelper::getMimeType($newfile), 'size' => filesize($newfile) ]);
+                        $attrValues[] = new UploadedFile(['name' => $thisFile['name'], 'tempName' => $newfile, 'type' => \yii\helpers\FileHelper::getMimeType($newfile), 'size' => filesize($newfile)]);
                      }
                      $this->owner->setAttribute($attr, $attrValues);
                   }
@@ -61,24 +61,24 @@ class UploadBehavior extends \yii\base\Behavior
    public function afterValidate()
    {
       foreach ($this->owner->rules() as $rule) {
-         if($rule[1] === 'file' || $rule[1] === 'image') {
+         if ($rule[1] === 'file' || $rule[1] === 'image') {
             if (is_array($rule[0])) {
                foreach ($rule[0] as $attr) {
                   $thisAttr = $this->owner->getAttribute($attr);
                   $attrValues = [];
-                  if($thisAttr !== null && is_array($thisAttr)) {
+                  if ($thisAttr !== null && is_array($thisAttr)) {
                      foreach ($thisAttr as $thisFile) {
                         $newFilename = '';
-                        for($i=0; $i<10; $i++) {
+                        for ($i = 0; $i < 10; $i++) {
                            $newFilename = trim($this->fieldsFolders[$attr], "/\\") . "/" . preg_replace("/([^A-Za-z0-9_])/", "_", \Yii::$app->security->generateRandomString(32)) . "." . $thisFile->extension;
-                           if(!file_exists(\Yii::getAlias('@app')."/../".$newFilename)) break;
+                           if (!file_exists(\Yii::getAlias('@app') . "/../" . $newFilename)) break;
                         }
                         $attrValues[] = $newFilename;
 
-                        if(count($this->owner->errors) > 0)
+                        if (count($this->owner->errors) > 0)
                            unlink($thisFile->tempName);
                         else
-                           rename($thisFile->tempName, \Yii::getAlias('@app')."/../".$newFilename);
+                           rename($thisFile->tempName, \Yii::getAlias('@app') . "/../" . $newFilename);
 
                      }
                      $this->owner->setAttribute($attr, implode(";", $attrValues));
@@ -89,8 +89,9 @@ class UploadBehavior extends \yii\base\Behavior
       }
    }
 
-   public function afterSave() {
-      if(count($this->owner->errors) > 0) {
+   public function afterSave()
+   {
+      if (count($this->owner->errors) > 0) {
          self::RemoveUnsavedFiles($this->owner);
       }
    }
@@ -99,18 +100,20 @@ class UploadBehavior extends \yii\base\Behavior
     * @param ActiveRecord $model
     * @throws \yii\base\Exception
     */
-   public static function RemoveUnsavedFiles($model) {
+   public static function RemoveUnsavedFiles($model)
+   {
 
       foreach ($model->rules() as $rule) {
-         if($rule[1] === 'file' || $rule[1] === 'image') {
+         if ($rule[1] === 'file' || $rule[1] === 'image') {
             if (is_array($rule[0])) {
                foreach ($rule[0] as $attr) {
                   $thisAttr = $model->getAttribute($attr);
                   $thisAttr = is_string($thisAttr) ? explode(";", $thisAttr) : [];
-                  if($thisAttr !== null && is_array($thisAttr)) {
+                  if ($thisAttr !== null && is_array($thisAttr)) {
                      foreach ($thisAttr as $thisFile) {
-                        $thisFile = \Yii::getAlias('@app')."/../".$thisFile;
-                        if(file_exists($thisFile)) {
+                        if ($thisFile === '') continue;
+                        $thisFile = \Yii::getAlias('@app') . "/../" . $thisFile;
+                        if (file_exists($thisFile)) {
                            unlink($thisFile);
                         }
                      }
